Fix bash env command (#2)
* Fix bash env command

* Add export to bash env command

`export` ensures that subshells inherit the new environment variable. Without it, users may encounter the following error message:
```
Unable to resolve service for type 'Microsoft.EntityFrameworkCore.Migrations.IMigrator'. This is often because no database provider has been configured for this DbContext. A provider can be configured by overriding the 'DbContext.OnConfiguring' method or by using 'AddDbContext' on the application service provider. If 'AddDbContext' is used, then also ensure that your DbContext type accepts a DbContextOptions<TContext> object in its constructor and passes it to the base constructor for DbContext.
```

This is because the environment specified doesn't make it through to the subprocess that is running this step and it attempts to use a configuration that may not have a configured SQL connection string.